updated ffmpeg process code

+++ GoDogSBPackage/GoDogSB.cs
@@ -1,7 +1,5 @@
 ﻿using System;
-using System.ComponentModel;
 using System.Diagnostics;
-using System.Threading;
 
 namespace GoDogSBPackage
 {
@@ -9,7 +7,6 @@ namespace GoDogSBPackage
     {
         protected static GoDogSB GoDog_SB;
         private GoDogLogger logger;
-        private Thread bgWorker;
         private Process process;
 
         public string InputURL { get; set; }
@@ -45,8 +42,7 @@ namespace GoDogSBPackage
         public void StartConversion()
         {
             IsForcedStopped = false;
-            bgWorker = new Thread(new ThreadStart(StartVideoConversion));
-            bgWorker.Start();
+            StartVideoConversion();
         }
 
         public void StopConversion(bool forcedStop = false)
@@ -54,11 +50,10 @@ namespace GoDogSBPackage
             try
             {
                 IsForcedStopped = forcedStop;
-                if (!process.HasExited)
+                if (process != null && !process.HasExited)
                 {
                     process.Kill();
                 }
-                bgWorker.Abort();
             }
             catch (Exception e)
             {
@@ -73,54 +68,64 @@ namespace GoDogSBPackage
 
         private void StartVideoConversion()
         {
-            if (string.IsNullOrWhiteSpace(this.InputURL) && string.IsNullOrWhiteSpace(this.OutputURL))
+            if (!string.IsNullOrWhiteSpace(this.InputURL) && !string.IsNullOrWhiteSpace(this.OutputURL))
             {
                 try
                 {
-                    string filArgs = string.Format("-stimeout 5000 -rtsp_transport tcp -i \"{0}\" -c:v copy -c:a aac -b:a 128k -ar 44100 -f flv \"{1}\"", this.InputURL, this.OutputURL);
+                    string filArgs = string.Format(" -rtsp_transport tcp -re -i \"{0}\" -c:v copy -c:a aac -b:a 128k -ar 44100 -f flv \"{1}\"", this.InputURL, this.OutputURL);
 
                     process = new Process();
+
                     process.StartInfo.FileName = AppDomain.CurrentDomain.BaseDirectory + "\\ffmpeg\\ffmpeg.exe";
                     process.StartInfo.Arguments = filArgs;
+
                     process.StartInfo.UseShellExecute = false;
-                    process.StartInfo.CreateNoWindow = false;
+                    process.StartInfo.CreateNoWindow = true;
                     process.StartInfo.RedirectStandardOutput = true;
                     process.StartInfo.RedirectStandardError = true;
+                    process.EnableRaisingEvents = true;
+
+                    process.Exited += Process_Exited;
+                    process.OutputDataReceived += Process_OutputDataReceived;
+                    process.ErrorDataReceived += Process_ErrorDataReceived;
 
                     process.Start();
-                    this.Log("Conversion process started.");
+                    process.BeginOutputReadLine();
+                    process.BeginErrorReadLine();
 
-                    while (!process.StandardError.EndOfStream)
-                    {
-                        this.Log(process.StandardError.ReadLine());
-                    }
+
+                    this.Log("Conversion process started.");
                 }
                 catch (Exception ex)
                 {
                     this.Log($"Exception: {ex.ToString()}");
-                }
-                finally
-                {
-                    if (process != null)
-                    {
-                        process.WaitForExit();
-                        process.Close();
-
-                        if (!IsForcedStopped)
-                        {
-                            this.Log("Restarting conversion process.");
-                            StopConversion();
-                            StartConversion();
-                        }
-
-                        this.Log("Conversion process exited.");
-                    }
                 }
             }
             else
             {
                 throw new Exception("Input/Output streams not found. PLease provoide Input/Output streams.");
             }
+        }
+
+        private void Process_OutputDataReceived(object sender, DataReceivedEventArgs e)
+        {
+            this.Log(e.Data);
+        }
+
+        private void Process_ErrorDataReceived(object sender, DataReceivedEventArgs e)
+        {
+            this.Log(e.Data);
+        }
+
+        private void Process_Exited(object sender, EventArgs e)
+        {
+            if (!IsForcedStopped)
+            {
+                this.Log("Restarting conversion process.");
+                StartConversion();
+            }
+
+            this.Log("Conversion process exited.");
         }
     }
 }
